jndklsah

+++ index.php
@@ -3,6 +3,15 @@ require_once('classes/database.php');
 $con = new database();
 session_start();
 
+if(!isset($_SESSION['username'])|| $_SESSION['account_type']){
+  header('location:login.php');
+  exit();
+}
+
+
+
+
+
 if (isset($_POST['delete'])) {
     $id = $_POST['id'];
     if ($con->delete($id)) {
@@ -117,6 +126,11 @@ document.addEventListener('DOMContentLoaded', function() {
       case 'success':
         title = 'Success!';
         text = 'Record is successfully deleted.';
+        icon = 'success';
+        break;
+      case 'success2':
+        title = 'Success!';
+        text = 'you are successfully logged in.';
         icon = 'success';
         break;
       case 'error':
